requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM & Embeddings
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
sentence-transformers>=2.2.0

# Vector Database
chromadb>=0.4.0

# Ollama Integration
ollama>=0.3.0

# OpenAPI Parsing
prance>=23.6.0
openapi-spec-validator>=0.7.0
PyYAML>=6.0

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
tenacity>=8.2.0
structlog>=23.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Cloud LLM fallback (uncomment if needed)
# groq>=0.4.0
